pyproject.toml
[project]
name = "strivee-btwb"
version = "0.1.0"
description = "Automate CrossFit programming transfer from Strivee to BTWB"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ollama>=0.4.0",
    "json-repair>=0.28.0",
    "Pillow>=11.0.0",
    "playwright>=1.44.0",
    "python-dotenv>=1.0.0",
    "pyobjc-framework-Quartz>=10.0",
]

[project.optional-dependencies]
dev = [
    "pytest==9.0.2",
    "pytest-cov==7.1.0",
    "pytest-watch==4.2.0",
    "ruff==0.9.2",
]

[project.scripts]
strivee-btwb = "strivee_btwb.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/strivee_btwb"]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "N",      # pep8-naming
    "UP",     # pyupgrade
    "PL",     # pylint
    "RUF",    # ruff-specific rules
]
ignore = [
    "E203",     # whitespace before ':'
    "PLR2004",  # Magic value used in comparison
    "PLR0913",  # Too many arguments in function definition (6 > 5)
]

[tool.ruff.lint.per-file-ignores]
"src/strivee_btwb/vision/parser.py" = ["E501"]  # prompt strings must not be line-wrapped
"src/strivee_btwb/processing/llm_format.py" = ["E501"]  # prompt strings must not be line-wrapped

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.coverage.run]
omit = ["src/strivee_btwb/__main__.py"]